+++ .eslintignore
@@ -1,2 +1,2 @@
-src/check.js
-src/*.min.js
+src/_not-used/check.js
+src/_not-used/*.min.js
