+++ src/filter/filter-type.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+    'ALL': 'filter-popular',
+    'DATE': 'filter-new',
+    'COMMENTS': 'filter-discussed'
+};
