+++ src/main.js
@@ -2,4 +2,4 @@
 
 require('./resizer');
 require('./upload');
-require('./pictures');
+require('./pictures/pictures');
